+++ CHANGELOG.md
@@ -25,6 +25,7 @@
       * h calculation does not fail if f(x) = 0
 * `__init__`
   * added `sgn` function to `scope`
+  * removed `Fraction` scope because of errors in `Function.derivative` and `integral`
 
 ## 3.1.0 (2022-01-01)